add option to delete existing files in branch (default true - backward c...

+++ src/main/groovy/org/ajoberstar/gradle/git/ghpages/GithubPagesPlugin.groovy
@@ -58,7 +58,7 @@ class GithubPagesPlugin implements Plugin<Project> {
 			def filesList = extension.workingDir.list({ dir, name ->
 				return !name.equals('.git')
 			})
-			if (filesList) {
+			if (filesList && extension.deleteExistingFiles) {
 				repo.remove(patterns: filesList)
 			}
 			project.copy {
